pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ultralight-frenzy"
version = "0.1.0"
description = "Agentic RAG app for ultralight backpacking gear"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "anthropic>=0.25",
    "chromadb>=0.5",
    "sentence-transformers>=3.0",
    "httpx>=0.27",
    "beautifulsoup4>=4.12",
    "typer>=0.12",
    "pydantic>=2.7",
    "pydantic-settings>=2.2",
    "python-dotenv>=1.0",
    "rich>=13",
]

[project.scripts]
gear = "cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["app", "agent", "db", "scrapers", "cli"]
